Enable automatic loading of CLI collections
