+++ cloud/aws/cli/plugins/cloudtrail.py
@@ -9,7 +9,7 @@ VALIDATORS = [
     dynaconf.Validator("VAST_CLOUDTRAIL_BUCKET_REGION", must_exist=True),
 ]
 
-TFDIR = "./terraform"
+INVOKE_CONFIG = {}
 
 
 @task
@@ -17,7 +17,7 @@ def deploy(c, auto_approve=False):
     """Deploy the cloudtrail datasource"""
     core.init_step(c, "cloudtrail")
     c.run(
-        f"terragrunt apply {core.auto_app_fmt(auto_approve)} --terragrunt-working-dir {TFDIR}/cloudtrail",
+        f"terragrunt apply {core.auto_app_fmt(auto_approve)} --terragrunt-working-dir {core.TFDIR}/cloudtrail",
         env=core.conf(VALIDATORS),
         pty=True,
     )
@@ -28,7 +28,7 @@ def destroy(c, auto_approve=False):
     """Remove the cloudtrail datasource"""
     core.init_step(c, "cloudtrail")
     c.run(
-        f"terragrunt destroy {core.auto_app_fmt(auto_approve)} --terragrunt-working-dir {TFDIR}/cloudtrail",
+        f"terragrunt destroy {core.auto_app_fmt(auto_approve)} --terragrunt-working-dir {core.TFDIR}/cloudtrail",
         env=core.conf(VALIDATORS),
         pty=True,
     )
